requirements.txt
# requirements.txt

# 主要なパッケージとバージョン
numpy==1.26.4
pandas==2.1.4
matplotlib==3.8.0
scipy==1.11.4
scikit-learn==1.2.2
jupyter==1.0.0
ipython==8.20.0
notebook==7.0.8
seaborn==0.12.2
statsmodels==0.14.0
plotly==5.9.0
bokeh==3.3.4
dask==2023.11.0
h5py==3.9.0
sqlalchemy==2.0.25
flask==2.2.5
beautifulsoup4==4.12.2
requests==2.31.0
pillow==10.2.0
networkx==3.1
sympy==1.12
pytest==7.4.0
sphinx==5.0.2
pyyaml==6.0.1
numba==0.59.1
nltk==3.8.1
spyder==5.4.3
streamlit==1.30.0
tabulate==0.9.0
tqdm==4.65.0
xlrd==2.0.1
openpyxl==3.0.10

# pip でインストールされたパッケージ
aiohttp-retry==2.8.3
amqp==5.2.0
antlr4-python3-runtime==4.9.3
asyncssh==2.14.2
atpublic==4.1.0
billiard==4.2.0
celery==5.4.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.8
dictdiffer==0.9.0
diskcache==5.6.3
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
efficientnet-pytorch==0.7.1
flatten-dict==0.4.2
flufl-lock==7.1.1
fsspec==2024.6.0
funcy==2.0
git-filter-repo==2.38.0
grandalf==0.8
gto==1.7.1
huggingface-hub==0.23.4
hydra-core==1.3.2
iterative-telemetry==0.0.8
kombu==5.3.7
munch==4.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orjson==3.10.5
pretrainedmodels==0.7.4
pydot==2.0.0
pygit2==1.15.0
pygtrie==2.5.0
safetensors==0.4.3
scmrepo==3.3.5
segmentation-models-pytorch==0.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
sqltrie==0.11.0
timm==0.9.2
torch==2.3.1
torchvision==0.18.1
triton==2.3.1
typer==0.12.3
vine==5.1.0
voluptuous==0.15.1
zc-lockfile==3.0.post1